remove taskId initialization and change type of id

+++ app/src/main/java/skills/future/planer/db/task/TaskDataDao.java
@@ -23,7 +23,7 @@ public interface TaskDataDao {
      * @throws Exception if sth is wrong xd
      */
     @Query("SELECT * FROM taskData WHERE taskDataId = :id")
-    TaskData findById(int id) throws Exception;
+    TaskData findById(Long id) throws Exception;
 
     /**
      * Method insert given taskData to database
